+++ src/main/java/com/project/doongdoong/domain/question/service/QuestionProvidable.java
@@ -1,14 +1,13 @@
 package com.project.doongdoong.domain.question.service;
 
 import com.project.doongdoong.domain.question.model.Question;
-import com.project.doongdoong.domain.question.model.QuestionContent;
 
 import java.util.List;
 
-public interface QuestionService {
+public interface QuestionProvidable {
 
     Question createFixedQuestion();
     Question createUnFixedQuestion();
 
-    List<Question> createQuestions();
+    List<Question> createRandomQuestions();
 }
